% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{Fhir browse locations}
\usage{
locations(state, suburb)
}
\arguments{
\item{state}{State}

\item{suburb}{Suburb}
}
\description{
This function returns all the centres
}
